+++ package.js
@@ -1,14 +1,14 @@
 /* global Package */
 Package.describe({
   name: 'freedombase:legal-management',
-  version: '2.0.0-rc.0',
+  version: '2.0.0-rc.1',
   summary: 'Manage your legal documents and user consent.',
   git: 'https://github.com/freedombase/meteor-legal-management',
   documentation: 'README.md'
 })
 
 Package.onUse(function (api) {
-  api.versionsFrom(['2.8.1', '3.0-beta.0'])
+  api.versionsFrom(['2.8.1', '3.0-rc.0'])
   api.use([
     'meteor',
     'ecmascript',
@@ -20,8 +20,8 @@ Package.onUse(function (api) {
     'zodern:types@1.0.11'
   ])
   api.use([
-    'aldeed:collection2@4.0.0',
-    'aldeed:schema-deny@4.0.0-beta.1',
+    'aldeed:collection2@4.0.1',
+    'aldeed:schema-deny@4.0.0',
     'aldeed:simple-schema@1.13.1',
   ])
   // 'socialize:base-model@1.1.2'