+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # CHANGELOG
 
-## v2.0.0 - 2024-02-XX
+## v2.0.0 - 2024-05-XX
 
 - Minimum Meteor version is now `2.8.1`, compatible with `v3.0-beta.0`
 - All collection calls have been migrated to async
 - Removed `aldeed:schema-index` as its functionality has been replaced by Meteor default `createIndexAsync` method.
-- Added `aldeed:simple-schame`
-- Upgraded `aldeed:collection2` to `v4.0.0`
+- Added `aldeed:simple-schema`
+- Upgraded `aldeed:collection2` to `v4.0.1`
 - Upgraded `aldeed:schema-deny` to `v4.0.0`
 
 ## v1.8.2 - 2023-12-30